% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_circle_plot.R
\name{draw_circle_plot}
\alias{draw_circle_plot}
\title{Draw a Plot of Circles}
\usage{
draw_circle_plot(numCirc)
}
\arguments{
\item{numCirc}{A vector of type numeric which defines the number of circles to draw}
}
\description{
Creates a blank plot with a border and then draws circles at random locations on the x and y axis. The circles each have a random color created using RGB values.
}
\examples{
draw_circle_plot(numCirc = 20)
}
