% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-words.R
\name{load_words}
\alias{load_words}
\title{Load words}
\usage{
load_words(wordLength)
}
\arguments{
\item{wordLength}{A character vector of either "easy", "medium", or "hard" which determines the number of syllables to use.}
}
\value{
A vector of character strings of length 5000 or 5865 representing English nouns.
}
\description{
Imports a file of every English nouns that are either one, two, or three syllables in length. In the case of two or three syllables, a sample of 5000 is chosen to prevent large vectors.
}
\examples{
load_words(wordLength = "easy")
load_words(wordLength = "hard")

}
\references{
\url{http://www.ashley-bovan.co.uk/words/partsofspeech.html}
}
